Takes into account error 404 while fetching NationBuilder
This correction solves a problem we have when a group or event is deleted
after it had been synchronized on our API. Whenever the import-rsvps
worker would try to fetch the RSVPs corresponding to our copy of the
event or group, NB would return a 404, given that it's not in NB anymore.

As I did not consider this case, fetchAll would have just tried again, ran out
of retries, and raised an error that would have triggered a restart of the
worker.

One such problem may thus completely block the sync of RSVPs.